[server & client] | refactor: PR 리뷰에서 지적된 사항 수정
